[Camera] Code clean
